Small correction of styling. Also, see description below:
Please find below a list of the files that have actually changed. All other changes are simply an
effect of moving the widely-used `Common` monad from `Node(Text)Editor.Action.Command`
to a common module: `Common.Action.Command`. Also, the imports formatting was
run, so the changes are almost exclusively import-related. The ones below you
should pay attention to, if you wish to review the code:

* luna-studio/lib/src/Common/Debug.hs
* luna-studio/node-editor/src/NodeEditor/Action/UUID.hs
* luna-studio/node-editor/src/NodeEditor/Handler/Backend/Common.hs
* luna-studio/node-editor/src/NodeEditor/Action/Command.hs ->
* luna-studio/lib/src/Common/Action/Command.hs
* luna-studio/node-editor/src/NodeEditor/State/Global.hs
* luna-studio/text-editor/src/TextEditor/Action/Batch.hs
* luna-studio/text-editor/src/TextEditor/Action/Command.hs (deleted)
* luna-studio/text-editor/src/TextEditor/Action/UUID.hs
* luna-studio/text-editor/src/TextEditor/State/Global.hs