Qualifind Data.Set imports with package name, so the build doesn't fail on Windows. (#1399)
Currently our packaging scripts fail to build runner on Windows:
```
[1 of 2] Compiling System.Host      ( src\System\Host.hs, src\System\Host.o )
[2 of 2] Compiling Main             ( src\StudioRunner.hs, src\StudioRunner.o )

src\StudioRunner.hs:17:1: error:
    Ambiguous interface for ‘Data.Set’:
      it was found in multiple packages:
      container-0.2.9 containers-0.5.7.1

src\StudioRunner.hs:28:1: error:
    Ambiguous interface for ‘Data.Set’:
      it was found in multiple packages:
      container-0.2.9 containers-0.5.7.1
```

The reason is that on Windows we call GHC directly to compile luna-studio.exe, rather than using stack. This is caused by the need to supply it with a custom pre-made resource binary blob, that adds icon to the generated executable (see #647). Apparently having a custom icon is not something that Haskell tooling supports.

This is obviously asking for trouble and this PR only alleviates the immediate issue… but it doesn't seem worthwhile to spend much time on figuring how this should be properly fixed (if possible at all), since the code is deprecated anyway.